final changes in refactoring
