+++ backend/workflow_manager/endpoint_v2/constants.py
@@ -2,6 +2,27 @@ import os
 from fnmatch import fnmatch
 
 
+class TableColumns:
+    CREATED_BY = "created_by"
+    CREATED_AT = "created_at"
+    METADATA = "metadata"
+    ERROR_MESSAGE = "error_message"
+    STATUS = "status"
+    USER_FIELD_1 = "user_field_1"
+    USER_FIELD_2 = "user_field_2"
+    USER_FIELD_3 = "user_field_3"
+    PERMANENT_COLUMNS = [
+        "created_by",
+        "created_at",
+        "metadata",
+        "error_message",
+        "status",
+        "user_field_1",
+        "user_field_2",
+        "user_field_3",
+    ]
+
+
 class DBConnectionClass:
     SNOWFLAKE = "SnowflakeDB"
     BIGQUERY = "BigQuery"
