+++ unstract/connectors/src/unstract/connectors/constants.py
@@ -54,24 +54,3 @@ class DatabaseTypeConstants:
     ORACLE_NUMBER: str = "NUMBER"
     ORACLE_LONG: str = "LONG"
     ORACLE_TIMESTAMP: str = "TIMESTAMP"
-
-
-class TableColumns:
-    CREATED_BY = "created_by"
-    CREATED_AT = "created_at"
-    METADATA = "metadata"
-    ERROR_MESSAGE = "error_message"
-    STATUS = "status"
-    USER_FIELD_1 = "user_field_1"
-    USER_FIELD_2 = "user_field_2"
-    USER_FIELD_3 = "user_field_3"
-    PERMANENT_COLUMNS = [
-        "created_by",
-        "created_at",
-        "metadata",
-        "error_message",
-        "status",
-        "user_field_1",
-        "user_field_2",
-        "user_field_3",
-    ]
